refactor: rename Sessions to History
